backend/requirements.txt
Flask==2.0.3
werkzeug==2.0.3
python-dotenv==0.19.0
plaid-python==9.1.0
flask-cors